Fix. SFW update. Inner SFW udpate functions moved to sfw-update-helper.php
